separate generator crate
