+++ crates/generator/src/client.rs
