+++ crates/generator/src/enums.rs
