+++ crates/generator/src/internal_enums.rs
