+++ crates/generator/src/models/types.rs
@@ -2,7 +2,7 @@ use prisma_client_rust_sdk::prisma::{
     prisma_models::walkers::ModelWalker, psl::parser_database::ScalarFieldType,
 };
 
-use crate::generator::prelude::*;
+use crate::prelude::*;
 
 fn scalar_selections_fn(model: ModelWalker, module_path: &TokenStream) -> TokenStream {
     let pcr = quote!(::prisma_client_rust);
