+++ crates/generator/src/models/with_params.rs
@@ -3,7 +3,7 @@ use prisma_client_rust_sdk::prisma::prisma_models::{
     FieldArity,
 };
 
-use crate::generator::prelude::*;
+use crate::prelude::*;
 
 use super::{order_by, pagination, ModelModulePart};
