+++ crates/generator/src/write_params.rs
